V2.0

+++ controllers/usuarios.controller.js
@@ -19,7 +19,7 @@ userCtrl.postUser = async (req, res) => {
             const user = new User({ Username, Email })
             await user.save()
 
-            // Send Message to Email --------------------------
+            // Send Message to Email ---------------------------
             const transporter = nodemailer.createTransport(
                 sendgridTransport({
                     auth: {
@@ -30,11 +30,13 @@ userCtrl.postUser = async (req, res) => {
             contentHTML = `
             <h1>Bienvenido a CUEFU</h1>
             <p>La mejor plataforma digital para escoger una universidad</p>
+            <div style="border-style: groove;">
             <h3>Tus Datos Personales</h3>
             <ul>
                 <li>Username: ${Username}</li>
                 <li>Correo Electrónico: ${Email}</li>
             </ul>
+            </div>
             <p>Muchas gracias por confiar en nosotros 
             <br> 
             En caso de una actualización te lo haremos saber</p>
@@ -50,7 +52,7 @@ userCtrl.postUser = async (req, res) => {
             transporter.sendMail(mailOptions, (error, info) => {
                 error ? console.log(error) : console.log(`Email Sent: ${info.response}`)
             })
-            // -----------------------------------------------
+            // ------------------------------------------------
             res.json({ status: 'User Saved and Email Sent' })
         }
     } catch (error) {
@@ -59,7 +61,47 @@ userCtrl.postUser = async (req, res) => {
 }
 
 userCtrl.deleteUser = async (req, res) => {
-    await User.findByIdAndDelete(req.params.id)
+    const { Username, Email } = req.body
+    try {
+        const emailUser = await User.findOne({ Email: Email })
+        if (!emailUser) {
+            // If emailUser already exist do something
+            throw new Error(`Email Doesn't Exist`)
+        } else {
+            await User.findByIdAndDelete(req.params.id)
+
+            // Send Message to Email ---------------------------
+            const transporter = nodemailer.createTransport(
+                sendgridTransport({
+                    auth: {
+                        api_key: process.env.SENDGRID_API_KEY
+                    }
+                }))
+
+            contentHTML = `
+            <h1>Mensaje CUEFU</h1>
+            <p>La mejor plataforma digital para escoger una universidad</p>
+            <div style="border-style: groove;" align="center">
+            <h3>Tu cuenta fue borrada Exitosamente</h3>
+            </div>
+            `
+
+            const mailOptions = {
+                from: process.env.EMAIL,
+                to: `${Email}`,
+                subject: "Cultura Universitaria - Escoge Tu Futuro",
+                html: `${contentHTML}`
+            }
+
+            transporter.sendMail(mailOptions, (error, info) => {
+                error ? console.log(error) : console.log(`Email Sent: ${info.response}`)
+            })
+            // ------------------------------------------------
+            res.json({ status: 'User Saved and Email Sent' })
+        }
+    } catch (error) {
+        console.log("ERROR >>> ", error)
+    }
     res.json({ status: 'User Deleted' })
 }
